Separate routes to different .js files
